+++ tests/test_reader.py
@@ -6,8 +6,8 @@ from pandas.testing import assert_frame_equal
 
 import tfs
 from tfs import read_tfs, write_tfs
-from tfs.errors import TfsFormatError
 from tfs.constants import HEADER
+from tfs.errors import TfsFormatError
 
 CURRENT_DIR = pathlib.Path(__file__).parent
 
@@ -34,6 +34,26 @@ class TestRead:
         assert len(test_file.index) > 0
         assert len(str(test_file)) > 0
         assert isinstance(test_file.index[0], str)
+
+    def test_tfs_read_no_validation(self, _tfs_file_pathlib: pathlib.Path):
+        test_file = read_tfs(_tfs_file_pathlib, index="NAME", validate=False)
+        assert len(test_file.headers) > 0
+        assert len(test_file.columns) > 0
+        assert len(test_file.index) > 0
+        assert len(str(test_file)) > 0
+        assert isinstance(test_file.index[0], str)
+
+    def test_tfs_read_wrong_file_no_validation(self, _space_in_colnames_tfs_path: pathlib.Path):
+        # Read file has a space in a column name which should raise, we make sure that it 
+        # goes through when explicitely skipping validation
+        df = read_tfs(_space_in_colnames_tfs_path, index="NAME", validate=False)
+        assert "BPM RES" in df.columns
+
+
+    def test_tfs_read_no_validation_doesnt_warn(self, caplog):
+        nan_tfs_path = pathlib.Path(__file__).parent / "inputs" / "has_nans.tfs"
+        _ = read_tfs(nan_tfs_path, index="NAME", validate=False)
+        assert "contains non-physical values at Index:" not in caplog.text
 
     def tfs_indx_pathlib_input(self, _tfs_file_pathlib: pathlib.Path):
         test_file = read_tfs(_tfs_file_pathlib)
@@ -104,10 +124,16 @@ class TestFailures:
         with pytest.raises(TfsFormatError):
             _ = tfs.reader._id_to_type(typoed_str_id)
 
+    def test_fail_space_in_colname(self, _space_in_colnames_tfs_path: pathlib.Path):
+        # Read file has a space in a column name which should raise
+        with pytest.raises(TfsFormatError):
+            read_tfs(_space_in_colnames_tfs_path, index="NAME", validate=True)
+
+
 class TestWarnings:
     def test_warn_unphysical_values(self, caplog):
         nan_tfs_path = pathlib.Path(__file__).parent / "inputs" / "has_nans.tfs"
-        _ = read_tfs(nan_tfs_path, index="NAME")
+        _ = read_tfs(nan_tfs_path, index="NAME", validate=True)
         for record in caplog.records:
             assert record.levelname == "WARNING"
         assert "contains non-physical values at Index:" in caplog.text
@@ -134,6 +160,11 @@ def _no_coltypes_tfs_path() -> pathlib.Path:
 @pytest.fixture()
 def _no_colnames_tfs_path() -> pathlib.Path:
     return pathlib.Path(__file__).parent / "inputs" / "no_colnames.tfs"
+
+
+@pytest.fixture()
+def _space_in_colnames_tfs_path() -> pathlib.Path:
+    return pathlib.Path(__file__).parent / "inputs" / "space_in_colname.tfs"
 
 
 @pytest.fixture()