+++ tfs/tools.py
@@ -1,6 +1,6 @@
 """
 Tools
------------------
+-----
 
 Additional functions to modify **TFS** files.
 """